@rynowak 2017-03-24
I have no perf numbers for this right now. I'm going to purse adding some microbenchmark style tests for this and then see what the delta is there as a first step.