@rynowak 2017-04-07
Rebased and added perf numbers:

**Before**
``` ini

BenchmarkDotNet=v0.10.3.0, OS=Microsoft Windows 10.0.15063
Processor=Intel(R) Core(TM) i7-3520M CPU 2.90GHz, ProcessorCount=4
Frequency=2825621 Hz, Resolution=353.9045 ns, Timer=TSC
dotnet cli version=2.0.0-preview1-005418
  [Host]     : .NET Core 4.6.25205.04, 64bit RyuJIT
  Job-PAWREA : .NET Core 4.6.25205.04, 64bit RyuJIT

RemoveOutliers=False  Runtime=Core  Server=True  
LaunchCount=3  RunStrategy=Throughput  TargetCount=10  
WarmupCount=5  

```
 |                                         Method |        Mean |    StdDev |     Op/s | Scaled | Scaled-StdDev | Allocated |
 |----------------------------------------------- |------------ |---------- |--------- |------- |-------------- |---------- |
 | 'conventional action selection implementation' |  14.2960 us | 0.2134 us | 69949.75 |   0.13 |          0.00 |   3.22 kB |
 |       'conventional action selection baseline' | 110.3249 us | 1.7803 us |  9064.14 |   1.00 |          0.00 |  62.02 kB |
```

**After**
``` ini

BenchmarkDotNet=v0.10.3.0, OS=Microsoft Windows 10.0.15063
Processor=Intel(R) Core(TM) i7-3520M CPU 2.90GHz, ProcessorCount=4
Frequency=2825621 Hz, Resolution=353.9045 ns, Timer=TSC
dotnet cli version=2.0.0-preview1-005418
  [Host]     : .NET Core 4.6.25205.04, 64bit RyuJIT
  Job-CLGBOK : .NET Core 4.6.25205.04, 64bit RyuJIT

RemoveOutliers=False  Runtime=Core  Server=True  
LaunchCount=3  RunStrategy=Throughput  TargetCount=10  
WarmupCount=5  

```
 |                                         Method |        Mean |    StdDev |     Op/s | Scaled | Scaled-StdDev | Allocated |
 |----------------------------------------------- |------------ |---------- |--------- |------- |-------------- |---------- |
 | 'conventional action selection implementation' |  13.0474 us | 0.1324 us | 76643.59 |   0.12 |          0.00 |   2.45 kB |
 |       'conventional action selection baseline' | 108.4359 us | 1.5474 us |  9222.04 |   1.00 |          0.00 |  62.02 kB |
```