@rynowak 2017-04-07
This is now ready for some intrepid explorer to give feedback on the design. I think this approach is much simpler than the big complicated decision tree.