@dougbu 2017-04-10
9 failures in CI builds look like real `ArgumentNullException`s