@rynowak 2017-04-21
Going to sleep on this until 2.0.0-preview2 as it's not a huge priority.