@rynowak 2017-05-05
@dougbu updated again. I also correctly named the test class, and deleted the dead code